// src/shared/types.ts
export type Id = string | number;

export interface Tag {
  id: Id;
  name: string;
  color: string;
}

export type UserRole = 'Admin' | 'Vendedor';

export interface User {
  id: string;
  name: string;
  email: string;
  avatarUrl?: string;
  role?: UserRole;
  joinedAt?: string;
}
// ...
export interface ColumnData {
  id: Id;
  title: string;
  color: string;
  type: 'open' | 'qualification' | 'follow-up' | 'scheduling' | 'won' | 'lost';
}